requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
torch==2.1.1
open_clip_torch==2.23.0
chromadb==0.4.18
numpy==1.26.2 